# perf: optimize OpenAPI parsing by reducing dict copy operations