# Description

  ## Problem Statement

  OpenAPI parsing performance regression causing excessive processing time for large schemas (e.g., GitHub API).
  Performance test failure: parsing time increased from expected <10s to 11.47s, exceeding the established threshold.

  ## Solution Overview

  Achieved significant performance improvement by optimizing dictionary copy operations in two critical functions that
  process OpenAPI schemas during parsing.

  🔧 Optimizations Implemented

  1. _make_optional_parameter_nullable Function
  - Before: Used schema.copy() to create full dictionary copy, then moved properties via del operations
  - After: Direct dictionary construction without intermediate copying and deletion
  - Enhancement: Replaced list iteration with set lookups for type-specific property classification

  2. _adjust_union_types Function
  - Before: Always copied input schema regardless of modification necessity
  - After: Added early-return logic to copy only when modifications are required
  - Enhancement: Smart detection of recursive processing needs to reduce unnecessary traversals

 ## 📊 Performance Results

  Test Case: GitHub API OpenAPI schema (~10MB)
  Before optimization: 11.47s  ❌ (exceeds 10s threshold)
  After optimization:   7.57s  ✅ (meets expectation)
  Improvement:         ~34% performance gain

  ## ✅ Quality Assurance

  - Functional integrity: All 142 OpenAPI-related tests pass
  - Behavioral consistency: Output identical to pre-optimization behavior
  - Type safety: Pyright type checking with 0 errors
  - Code standards: Passes all pre-commit checks (ruff, prettier, pyright)

  
  ## Technical Details

  - Memory efficiency: Reduced memory allocations during schema processing
  - Algorithm improvement: O(n) single-pass property distribution vs O(n²) copy-and-delete
  - Maintainability: Clearer code structure with explicit property handling logic