@chi2liu 2025-08-01
> Do you mean that the unit test is failing on your machine? If you run `pytest tests/experimental/server/test_openapi_performance.py::TestOpenAPIPerformance::test_github_api_schema_performance -s` you'll see a print at the end with how long the OpenAPI parsing takes for the entire GitHub API. You're seeing that exceed 10s?

Yes, the unit test is failing on my machine.  However, I believe this is primarily an environment performance
  issue rather than a problem with the code themselves. The optimizations from the PR should still provide the same ~34% performance improvement mentioned in the commit message, and the results are within the expected range given environment variations.